added loading of collision models to workcell gl drawer
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@2040 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rwlibs/drawable/ac3d/LoaderAC3D.cpp
@@ -212,7 +212,7 @@ Model3DPtr LoaderAC3D::load(const std::string& filename){
                     std::vector<int> indices;
                     int iidx=0;
                     if( Triangulate::processPoints(points, indices) ){
-                        while(iidx<indices.size()){
+                        while(iidx < (int)indices.size()){
                             IndexedTriangle<> tri(poly[ indices[iidx  ] ],
                                                   poly[ indices[iidx+1] ],
                                                   poly[ indices[iidx+2] ]);
